+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- hnefatafl-client: automatically resize the screen.
 - hnefatafl-client: don't mention you're leaving unless you're the attacker or defender.
 
 ## [0.11.0] - 2025-03-19